Remove nested spec method
It was previously being used for convenience during testing of Uspec itself, but it created a lot of weird situations.

By making the Spec object it meant that there would be nested specs but they could behave differently than top-level specs when it comes to environment.

This probably will be re-added in the future, but with better support so that nested specs fail their parents and display properly in output.